move parsing stuff to `core::parser.rs`
instead of doing all the parsing in hvm.rs, move it to a separate
file
