+++ kindelia/tests/cli.rs
@@ -7,6 +7,7 @@ mod cli {
   use kindelia_core::api;
   use kindelia_core::common;
   use kindelia_core::hvm;
+  use kindelia_core::parser;
 
   const CRATE_NAME: &str = assert_cmd::crate_name!();
 
@@ -53,7 +54,7 @@ mod cli {
 
     // read file and get statments
     let file_content = std::fs::read_to_string(file).unwrap();
-    let (_, s1) = hvm::read_statements(&file_content).unwrap();
+    let (_, s1) = parser::read_statements(&file_content).unwrap();
 
     // serializes file and saves it in a temp file
     let output = kindelia!().args(["serialize", file]).output().unwrap();
@@ -72,7 +73,7 @@ mod cli {
       .output()
       .unwrap();
     let output = get_stdout(&output);
-    let (_, s2) = hvm::read_statements(&output).unwrap();
+    let (_, s2) = parser::read_statements(&output).unwrap();
 
     eprintln!("OUT:\n{}", output);
 
@@ -107,7 +108,7 @@ mod cli {
     use kindelia_core::bits::ProtoSerialize;
     eprintln!("{}", file);
     let txt = std::fs::read_to_string(file).unwrap();
-    let stmts = hvm::parse_code(&txt).unwrap();
+    let stmts = parser::parse_code(&txt).unwrap();
     let str_0 = hvm::view_statements(&stmts);
     let ser = stmts.proto_serialized();
     let deser = Vec::proto_deserialized(&ser).unwrap();
@@ -254,7 +255,7 @@ mod cli {
 
   fn fun_response_1() -> (api::FuncInfo, Option<hvm::Term>) {
     let code = FUN_CODE;
-    let func = hvm::read_statement(code).unwrap().1;
+    let func = parser::read_statement(code).unwrap().1;
     if let hvm::Statement::Fun { func, init, .. } = func {
       (api::FuncInfo { func }, init)
     } else {
